apps/sovereign/scripts/sovereign_dag/requirements.txt
pydantic>=2.0.0
psycopg2-binary
requests
numpy
python-dotenv
instructor
ollama
